backend/requirements.txt
python-dotenv
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
pillow
dj_database_url
cloudinary
django-cloudinary-storage